Allow configuring default newline separator
