+++ Bonsai.System/IO/Ports/ObservableSerialPort.cs
@@ -1,47 +1,11 @@
 ﻿using System;
 using System.Reactive.Linq;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace Bonsai.IO
 {
     static class ObservableSerialPort
     {
-        public const string DefaultNewLine = @"\r\n";
-
-        public static string Unescape(string value)
-        {
-            return Regex.Replace(value, "\\\\[\'\"\\\\0abfnrtv]|\\\\u[0-9a-fA-F]{4}|\\\\U[0-9a-fA-F]{8}|\\\\x[0-9a-fA-F]{0,4}", m =>
-            {
-                if (m.Length == 1) return m.Value;
-                if (m.Value[1] == 'u' || m.Value[1] == 'x')
-                {
-                    return new string((char)Convert.ToInt32(m.Value.Substring(2), 16), 1);
-                }
-                if (m.Value[1] == 'U')
-                {
-                    var utf32 = Convert.ToInt32(m.Value.Substring(2), 16);
-                    return char.ConvertFromUtf32(utf32);
-                }
-
-                switch (m.Value)
-                {
-                    case @"\'": return "\'";
-                    case @"\""": return "\"";
-                    case @"\\": return "\\";
-                    case @"\0": return "\0";
-                    case @"\a": return "\a";
-                    case @"\b": return "\b";
-                    case @"\f": return "\f";
-                    case @"\n": return "\n";
-                    case @"\r": return "\r";
-                    case @"\t": return "\t";
-                    case @"\v": return "\v";
-                    default: return m.Value;
-                }
-            });
-        }
-
         public static IObservable<string> ReadLine(string portName, string newLine)
         {
             return Observable.Create<string>((observer, cancellationToken) =>
@@ -52,6 +16,7 @@ namespace Bonsai.IO
                     using var connection = SerialPortManager.ReserveConnection(portName);
                     using var cancellation = cancellationToken.Register(connection.Dispose);
                     var serialPort = connection.SerialPort;
+                    if (string.IsNullOrEmpty(newLine)) newLine = serialPort.NewLine;
                     while (!cancellationToken.IsCancellationRequested)
                     {
                         try
@@ -87,6 +52,24 @@ namespace Bonsai.IO
                 TaskCreationOptions.LongRunning,
                 TaskScheduler.Default);
             });
+        }
+
+        public static IObservable<TSource> WriteLine<TSource>(IObservable<TSource> source, string portName, string newLine)
+        {
+            return Observable.Using(
+                () => SerialPortManager.ReserveConnection(portName),
+                connection =>
+                {
+                    if (string.IsNullOrEmpty(newLine)) newLine = connection.SerialPort.NewLine;
+                    return source.Do(value =>
+                    {
+                        lock (connection.SerialPort)
+                        {
+                            connection.SerialPort.Write(value.ToString());
+                            connection.SerialPort.Write(newLine);
+                        }
+                    });
+                });
         }
     }
 }
