+++ Bonsai.System/IO/Ports/SerialPortManager.cs
@@ -6,11 +6,13 @@ using System.Xml;
 using System.IO;
 using System.Reactive.Disposables;
 using System.IO.Ports;
+using System.Text.RegularExpressions;
 
 namespace Bonsai.IO
 {
     internal static class SerialPortManager
     {
+        public const string DefaultNewLine = @"\r\n";
         public const string DefaultConfigurationFile = "SerialPort.config";
         static readonly bool IsRunningOnMono = Type.GetType("Mono.Runtime") != null;
         static readonly Dictionary<string, Tuple<SerialPort, RefCountDisposable>> openConnections = new();
@@ -67,6 +69,13 @@ namespace Bonsai.IO
                     {
                         serialPort.Encoding = Encoding.GetEncoding(encoding);
                     }
+
+                    var newLine = serialPortConfiguration.NewLine;
+                    if (!string.IsNullOrEmpty(newLine))
+                    {
+                        serialPort.NewLine = Unescape(newLine);
+                    }
+
                     serialPort.Open();
 
                     if (serialPort.BytesToRead > 0)
@@ -87,6 +96,39 @@ namespace Bonsai.IO
 
                 return new SerialPortDisposable(connection.Item1, connection.Item2.GetDisposable());
             }
+        }
+
+        public static string Unescape(string value)
+        {
+            return Regex.Replace(value, "\\\\[\'\"\\\\0abfnrtv]|\\\\u[0-9a-fA-F]{4}|\\\\U[0-9a-fA-F]{8}|\\\\x[0-9a-fA-F]{0,4}", m =>
+            {
+                if (m.Length == 1) return m.Value;
+                if (m.Value[1] == 'u' || m.Value[1] == 'x')
+                {
+                    return new string((char)Convert.ToInt32(m.Value.Substring(2), 16), 1);
+                }
+                if (m.Value[1] == 'U')
+                {
+                    var utf32 = Convert.ToInt32(m.Value.Substring(2), 16);
+                    return char.ConvertFromUtf32(utf32);
+                }
+
+                switch (m.Value)
+                {
+                    case @"\'": return "\'";
+                    case @"\""": return "\"";
+                    case @"\\": return "\\";
+                    case @"\0": return "\0";
+                    case @"\a": return "\a";
+                    case @"\b": return "\b";
+                    case @"\f": return "\f";
+                    case @"\n": return "\n";
+                    case @"\r": return "\r";
+                    case @"\t": return "\t";
+                    case @"\v": return "\v";
+                    default: return m.Value;
+                }
+            });
         }
 
         [Obsolete]
